[Notifier][FrameworkBundle] Add NotifierBundle
The notifier services move to the Notifier component, which now provides its own bundle and
the "notifier" configuration. The "framework.notifier" configuration becomes an alias of it.
The transport factories of every notifier bridge move with them, so the list of bridges lives
next to the bridges themselves instead of in FrameworkExtension.

A new RemoveMissingDependenciesPass takes over every decision the extension used to make by
reading another section, which also makes the bundle usable on its own: the channels drop
their transport when a default message bus is registered, the email channel takes the mailer
envelope sender or is dropped when no mailer is around, the data collector is registered in
debug mode and dropped when the profiler is disabled, and the notification logger listener is
gated on the profiler state, or dropped when neither the profiler nor the test client
consumes the notifications it keeps.

That pass absorbs RemoveMissingMailerDependenciesPass, which existed in FrameworkBundle only
to wire the email channel, and the notifier half of DefaultMessageBusPass.
