--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/notifier.php
@@ -1,27 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'messenger' => [
-        'enabled' => true,
-    ],
-    'mailer' => [
-        'dsn' => 'smtp://example.com',
-    ],
-    'notifier' => [
-        'enabled' => true,
-        'notification_on_failed_messages' => true,
-        'chatter_transports' => [
-            'slack' => 'null',
-        ],
-        'texter_transports' => [
-            'twilio' => 'null',
-        ],
-        'channel_policy' => [
-            'low' => ['slack'],
-            'high' => ['slack', 'twilio'],
-        ],
-        'admin_recipients' => [
-            ['email' => 'test@test.de', 'phone' => '+490815',],
-        ],
-    ],
-]);
