--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/notifier_with_specific_message_bus.php
@@ -1,19 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'messenger' => [
-        'enabled' => true,
-    ],
-    'mailer' => [
-        'dsn' => 'smtp://example.com',
-    ],
-    'notifier' => [
-        'message_bus' => 'app.another_bus',
-        'chatter_transports' => [
-            'test' => 'null',
-        ],
-        'texter_transports' => [
-            'test' => 'null',
-        ],
-    ],
-]);
