--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/notifier_without_messenger.php
@@ -1,28 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'mailer' => [
-        'dsn' => 'smtp://example.com',
-    ],
-    'messenger' => [
-        'enabled' => false,
-    ],
-    'notifier' => [
-        'enabled' => true,
-        'notification_on_failed_messages' => true,
-        'chatter_transports' => [
-            'slack' => 'null',
-        ],
-        'texter_transports' => [
-            'twilio' => 'null',
-        ],
-        'channel_policy' => [
-            'low' => ['slack'],
-            'high' => ['slack', 'twilio'],
-        ],
-        'admin_recipients' => [
-            ['email' => 'test@test.de', 'phone' => '+490815',],
-        ],
-    ],
-    'scheduler' => false,
-]);
