--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/notifier_without_transports.php
@@ -1,7 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'notifier' => [
-        'enabled' => true,
-    ],
-]);
