+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/legacy_notifier.yml
@@ -1,0 +1,4 @@
+framework:
+    notifier:
+        texter_transports:
+            twilio: 'null'
