--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/notifier.yml
@@ -1,17 +1,0 @@
-framework:
-    messenger:
-        enabled: true
-    mailer:
-        dsn: 'smtp://example.com'
-    notifier:
-        enabled: true
-        notification_on_failed_messages: true
-        chatter_transports:
-            slack: 'null'
-        texter_transports:
-            twilio: 'null'
-        channel_policy:
-            low: ['slack']
-            high: ['slack', 'twilio']
-        admin_recipients:
-            - { email: 'test@test.de', phone: '+490815' }
