--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/notifier_with_specific_message_bus.yml
@@ -1,11 +1,0 @@
-framework:
-    messenger:
-        enabled: true
-    mailer:
-        dsn: 'smtp://example.com'
-    notifier:
-        message_bus: 'app.another_bus'
-        chatter_transports:
-            test: 'null'
-        texter_transports:
-            test: 'null'
