--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/notifier_without_messenger.yml
@@ -1,18 +1,0 @@
-framework:
-    mailer:
-        dsn: 'smtp://example.com'
-    messenger:
-        enabled: false
-    notifier:
-        enabled: true
-        notification_on_failed_messages: true
-        chatter_transports:
-            slack: 'null'
-        texter_transports:
-            twilio: 'null'
-        channel_policy:
-            low: ['slack']
-            high: ['slack', 'twilio']
-        admin_recipients:
-            - { email: 'test@test.de', phone: '+490815' }
-    scheduler: false
