--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/notifier_without_transports.yml
@@ -1,3 +1,0 @@
-framework:
-    notifier:
-        enabled: true
