+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -59,7 +59,7 @@
         "symfony/mailer": "^8.2",
         "symfony/messenger": "^8.2",
         "symfony/mime": "^8.2",
-        "symfony/notifier": "^7.4|^8.0",
+        "symfony/notifier": "^8.2",
         "symfony/object-mapper": "^8.2",
         "symfony/polyfill-intl-icu": "^1.0",
         "symfony/process": "^8.2",
@@ -97,6 +97,7 @@
         "symfony/json-streamer": "<8.2",
         "symfony/mailer": "<8.2",
         "symfony/lock": "<8.2",
+        "symfony/notifier": "<8.2",
         "symfony/security-csrf": "<7.4",
         "symfony/semaphore": "<8.2",
         "symfony/serializer": "<7.4",
