+++ src/Symfony/Component/Notifier/Resources/config/notifier.php
