+++ src/Symfony/Component/Notifier/composer.json
@@ -21,6 +21,8 @@
         "symfony/deprecation-contracts": "^2.5|^3"
     },
     "require-dev": {
+        "symfony/config": "^7.4|^8.0",
+        "symfony/dependency-injection": "^8.2",
         "symfony/event-dispatcher-contracts": "^2.5|^3",
         "symfony/event-dispatcher": "^6.4|^7.0|^8.0",
         "symfony/http-client-contracts": "^2.5|^3",
@@ -29,6 +31,7 @@
         "symfony/messenger": "^7.4|^8.0"
     },
     "conflict": {
+        "symfony/dependency-injection": "<8.2",
         "symfony/event-dispatcher-contracts": "<2.5",
         "symfony/http-client-contracts": "<2.5"
     },
